manual: Create empty placeholder macros for @standards.
Header and standards annotations are slated for standardization,
including being rendered in the description of functions, variables,
etc. (elements), and eventually required.  This commit adds @standards
dummy macros so we can convert all existing annotations to the new
framework while maintaining the rendered status quo.

There needs to be a way to disambiguate annotations in lists of @*x
elements, where a common description is shared but some elements may
have different headers or standards.  The @standardsx macro fills this
role by accepting an additional parameter: the name of the annotated
element.

	* manual/macros.texi (@standards): New macro.  Provide
	placeholder for header and standards annotations.
	(@standardsx): New macro.  Likewise, for lists of @*x
	elements.