Fix signal.h bsd_signal namespace (bug 21445).
signal.h declares bsd_signal for __USE_XOPEN.  But this function was
obsoleted in the 2001 edition of POSIX and removed in the 2008
edition, so it should not be declared when the 2008 edition is in use.
This patch fixes the conditionals accordingly.  (This does not fix any
conform/ test failures because of other namespace issues in signal.h.)

Tested for x86_64.

	[BZ #21445]
	* signal/signal.h [__USE_XOPEN2K8] (bsd_signal): Do not declare.